Add info message on page.

+++ Controllers/UserController.cs
@@ -35,19 +35,22 @@ namespace LibraryManagement.Controllers
             User user = userService.Register(login, password, password_repeat);
             if (user.login is not null)
             {
+                TempData["Message"] = "Zostałeś zarejestrowany.";
                 HttpContext.Session.SetString("login", login);
                 HttpContext.Session.SetString("isSuperUser", user.isSuperUser ? "true" : "false");
-                return RedirectToAction(nameof(Welcome));
+
+                return RedirectToAction(nameof(Index));
             }
             else
             {
-                ViewBag.msg = "Invalid";
+                TempData["MessageErr"] = "Rejestracja nie powiodła się.";
                 return RedirectToAction(nameof(HomeController.Index), "Home");
             }
         }
         public ActionResult Logout()
         {
             HttpContext.Session.Clear();
+            TempData["Message"] = "Zostałeś wylogowany.";
             return RedirectToAction(nameof(HomeController.Index), "Home");
         }
         public ActionResult Login()
@@ -61,12 +64,13 @@ namespace LibraryManagement.Controllers
             User user = userService.Login(login, password);
             if (user is not null)
             {
+                TempData["Message"] = "Zostałeś zalogowany.";
                 HttpContext.Session.SetString("login", login);
                 HttpContext.Session.SetString("isSuperUser", user.isSuperUser ? "true" : "false");
             }
             else
             {
-                ViewBag.msg = "Invalid";
+                TempData["MessageErr"] = "Logowanie nie powiodło się.";
             }
             return RedirectToAction(nameof(HomeController.Index), "Home");
         }
@@ -82,15 +86,24 @@ namespace LibraryManagement.Controllers
             string? login = HttpContext.Session.GetString("login");
             if (!string.IsNullOrEmpty(login))
             {
-                userRepository.Delete(userRepository.GetUser(login).UserID);
-                HttpContext.Session.Clear();
+                User user = userRepository.GetUser(login);
+                if (user.books.Count() == 0)
+                {
+                    TempData["Message"] = "Konto zostało usunięte.";
+                    userRepository.Delete(userRepository.GetUser(login).UserID);
+                    HttpContext.Session.Clear();
+                }
+                else
+                {
+                    TempData["MessageErr"] = "Nie można usunąć konta, bo są do niego przypisane książki.";
+                }
+            }
+            else
+            {
+                TempData["MessageErr"] = "Nie udało się usunąć konta.";
             }
             return RedirectToAction(nameof(HomeController.Index), "Home");
  
-        }
-        public ActionResult Welcome()
-        {
-            return View();
         }
